Vehicle marker Duplication fix
The if statment for whether to update a marker or not would first check the marker exists and then check if the marker's current location on the map. I have now changed this to a child if statement which now means the change in location will be animated if the previous location is on the map, as was previously intended.